% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcolor.R
\name{mdcolor}
\alias{mdcolor}
\title{Font and Text background Color in Markdown}
\usage{
mdcolor()
}
\value{
character code for including the \code{mdcolor.js}
        JavaScript library
}
\description{
Extending the Markdown markup language with syntax for changing
             the font color and the text background color in the document.
             the R package is based on a JavaScript
             library called \code{mdcolor.js}.
}
