% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcolor.R
\name{mdcolor}
\alias{mdcolor}
\title{Font and Text background Color in Markdown}
\usage{
mdcolor()
}
\value{
HTML code for sourcing the \code{mdcolor.js}
        JavaScript library in the Rmarkdown file
}
\description{
Extending the Markdown markup language with
             syntax for changing
             the font color and the text background color
             in the document.
             the R package is based on a JavaScript
             library called \code{mdcolor.js}. for syntax
             explanation, see the details section below.
}
\details{
All of R colors are supported by \code{mdcolor} package.
you can see the name of R colors by typing \code{color()}
function. the syntax for applying a color to text or
text background are shown below:

\code{    [colorname]} change text color \code{[#]} \cr
\code{    [-colorname]} highlight text \code{[#]} \cr

as noted above, the syntax is only different in a dash that
is placed before the color name, inorder to specify the
background color.
}
\examples{
\dontrun{
# include the following code chunks in an Rmd file
# ================================================

```{r, echo=FALSE, results='asis'}
library(mdcolor)
mdcolor()
```

[green]Fun Begins![#]
---------------------

This is an R [-yellow] Markdown document[#].
[-purple]Markdown[#] is a simple formatting syntax for
authoring [green]HTML[#], [blue]PDF[#], and [red]MS[#]
Word documents. As you can see, All [-hotpink]R[#]
[orangered]colors[#]are
[-greenyellow][orange]**AVAILABL**[#][#] and the syntax
can be used with Markdown.

**GitHub Repository**:  https://github.com/haghish/mdcolor
}

}
